refactor(admin): Q9.f — extract admin/secrets.rs (GAR-475)
Moves all secrets-management code (CRUD, rotation, migration,
AES-256-GCM helpers, and config-secret redaction) from the 1738-LOC
admin/handlers.rs into a new focused module admin/secrets.rs.

handlers.rs shrinks 1738 → 1270 LOC (−468 lines). Re-exports in
handlers.rs preserve all routes.rs call-sites unchanged. Zero
behaviour change; only code organisation changes.

Security invariants preserved:
- encrypt_value / decrypt_value remain private in secrets.rs
- test_secret returns only value_length (int), never decrypted bytes
- redact_config_secrets masks all api_key / token / secret / password
  fields and is pub(super) — callable from config handlers in handlers.rs
- No new log paths expose encryption_key bytes
