+++ crates/garraia-gateway/src/admin/mod.rs
@@ -7,6 +7,7 @@ pub mod observability;
 pub mod providers;
 pub mod rbac;
 pub mod routes;
+pub mod secrets;
 pub mod shared;
 pub mod store;
 pub mod users;
